refactor(holdings): extract HoldingProperties component, add page padding
- Move all holding edit/create logic from HoldingDetailPage into a new
  HoldingProperties component (src/client/components/HoldingProperties).
  Component name now matches the rendered div class — same convention as
  TransactionProperties / ConnectionProperties.
- HoldingDetailPage is now a thin wrapper:
    <div className="HoldingDetailPage"><HoldingProperties /></div>
- Add canonical page padding: div.HoldingDetailPage > div { padding: 0 10px; }
  matching AccountDetailPage / TransactionDetailPage / ConnectionDetailPage.

Addresses hoiekim review (May 8) on PR #196.
