+++ src/client/components/index.ts
@@ -1,4 +1,5 @@
 export * from "./App";
+export * from "./HoldingProperties";
 export * from "./HoldingsManager";
 export * from "./ErrorBoundary";
 export * from "./PlaidLinkContext";
